Crate portable_atomic_util

Synchronization primitives built with portable-atomic.

Provide Arc. (optional, requires the std or alloc feature)

Provide task::Wake. (optional, requires the std or alloc feature)

See portable-atomic#1 for other primitives being considered for addition to this crate.

This crate was originally part of the portable-atomic repository and was extracted into its own repository.

Optional features

std
Use std.

Note:

This implicitly enables the alloc feature.

alloc
Use alloc.

Note:

The MSRV when this feature is enabled and the std feature is not enabled is Rust 1.36 that alloc crate stabilized.

serde
Implement serde::{Serialize, Deserialize} for Arc.

Note:

The MSRV when this feature is enabled is the one coming from serde, as of now it’s Rust 1.56

Optional cfg

One of the ways to enable cfg is to set rustflags in the cargo config:

# .cargo/config.toml
[target.<target>]
rustflags = ["--cfg", "portable_atomic_unstable_coerce_unsized"]

Or set environment variable:

RUSTFLAGS="--cfg portable_atomic_unstable_coerce_unsized" cargo ...

--cfg portable_atomic_unstable_coerce_unsized
Support coercing of Arc<T> to Arc<U> as in std::sync::Arc.

This cfg requires Rust nightly because this coercing requires unstable CoerceUnsized trait.

See this issue comment for another known workaround.

Note: This cfg is unstable and outside of the normal semver guarantees and minor or patch versions of portable-atomic-util may make breaking changes to them at any time.

Related Projects

portable-atomic: Portable atomic types including support for 128-bit atomics, atomic float, etc.

atomic-maybe-uninit: Atomic operations on potentially uninitialized integers.

atomic-memcpy: Byte-wise atomic memcpy.

Modules

task

Types and Traits for working with asynchronous tasks.

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Weak

Weak is a version of Arc that holds a non-owning reference to the managed allocation.